converted int to float for price in Offer

+++ Core/src/main/java/ir/sharif/gamein2021/core/domain/dto/OfferDto.java
@@ -17,7 +17,7 @@ public class OfferDto implements BaseDto<Integer>
     private Integer productId;
     private Integer volume;
     private OfferStatus offerStatus;
-    private Integer costPerUnit;
+    private Float costPerUnit;
     private LocalDate offerDeadline;
 
     @Override
